refactor: dedup _ensure_neighbor_future_4col, require wheelbase in _ego_obb_corners
- Move _ensure_neighbor_future_4col into scene_render.py (gradio-free, shared) and
  import it in scene_branch_editor and viz_saved_scenes instead of two copies.
- recovery_sim._ego_obb_corners: make wheelbase a required arg (drop the
  length*0.65 silent fallback; the sole caller already passes ego_wheelbase).
- generate_samples: keep make_initial_latent import at function scope (module-top
  import would create an import cycle via batched_rollout); hoisted out of the loop.
